==> 01_Data Preparation/01_Data_Preparation.ipynb <==
# Data Preparation

Drop unimportant data to decrease file size for uploading it on Colab.

import pandas as pd

data = pd.read_csv("../data/01_Original.csv")
data.head(3)

The Data Exploration shows why the author_email was chosen over the committer_email.

data.drop(['hash', 'diff', 'committer_email', 'author_name', 'committer_name', 'split'], axis=1, inplace=True)
data

data.to_pickle('../data/02_All_Decreased_Filesize.pkl')
