==> 01_Data Preparation/02_Frequent_Committer_Subset.ipynb <==
# Generate Most Frequent Committers Subset

Goal: Include all messages from committers who have more than 1000 commits.

#### Load Data

import pandas as pd

data = pd.read_csv("../data/01_Original.csv")
data.head(3)

data.drop(['hash', 'diff', 'committer_email', 'author_name', 'committer_name', 'split'], axis=1, inplace=True)

#### Take the subset

The Dataset Exploration showed that there are exactly 42 committers who have more than 1000 commits taken by author_email.

from collections import Counter

author_count = Counter(data["author_email"])
included_authors_counter = author_count.most_common(42)

author_count.most_common(43)

#### Transform to a list

included_author_emails = []

for tuple in included_authors_counter:
    if tuple[1] >= 1000: #not really required since already filtered above
        included_author_emails.append(tuple[0])

included_author_emails

#### Filter the dataframe

subset = data.where(data['author_email'] == included_author_emails[0])

for author_email in included_author_emails[1:]:
    subset = pd.concat([subset, data.where(data['author_email'] == author_email)])

subset = subset.dropna()
subset.reset_index(drop=True, inplace=True)
subset

#### Check whether the sum is correct

sum([count for _, count in included_authors_counter])

The sum is correct.

#### Save dataset

subset.to_pickle('../data/03_Subset_Frequent_Committers.pkl')
subset.to_csv('../data/03_Subset_Frequent_Committers.csv')
