==> 01_Data Preparation/02b_Projects_Subset.ipynb <==
# Generate Most Frequent Projects Subset

Goal: Include all messages from the 100 projects that have the most commit messages.

#### Load Data

import pandas as pd

data = pd.read_pickle("../data/02_All_Decreased_Filesize.pkl")
data.head(3)

#### Take the subset

For the final model, the 100 projects with the highest amount of commit messages are used.

from collections import Counter

project_count = Counter(data["project"])
included_projects_counter = project_count.most_common(100)

project_count.most_common(100)

#### Transform to a list

included_projects = [project[0] for project in included_projects_counter]
included_projects

#### Filter the dataframe

subset = data.where(data['project'] == included_projects[0])

for project in included_projects[1:]:
    subset = pd.concat([subset, data.where(data['project'] == project)])

subset = subset.dropna()
subset.reset_index(drop=True, inplace=True)
subset

#### Check whether the sum is correct

sum([count for _, count in included_projects_counter])

The sum is correct.

#### Save dataset

subset.to_pickle('../data/03b_Projects_Subset.pkl')
