==> 01_Data Preparation/03_Train_Val_Test_Split.ipynb <==
# Split the Frequent Committers Dataset

Goal: Split the Data with the most frequent committers into a train (70 %), validate (15 %) and test set (15 %).

#### Load Data

import pandas as pd

data = pd.read_pickle('../data/03_Subset_Frequent_Committers.pkl')
data.head(3)

from collections import Counter

author_counter = Counter(data['author_email'])
len(author_counter)

There are 42 authors.

All messages of one author should be in one subset to be able to cluster authors by their style.

print('70 percent of the {total_authors} authors: {fraction_authors} authors'.format(total_authors = len(author_counter), fraction_authors = len(author_counter) * 0.7))
print('15 percent of the {total_authors} authors: {fraction_authors} authors'.format(total_authors = len(author_counter), fraction_authors = len(author_counter) * 0.15))

Since these numbers are really small, a larger validate and test subset is preffered.

The following split is proposed:

| | number of authors |
| --- | --- |
| Training | 28 |
| Validate | 7 |
| Test | 7 |

The allocation of authors to a subset should be random because some authors have more commit messages while others have fewer.

import numpy as np

# this seed (13) returns a splitting option that fullfills the above splitting proportionalities, other most probably may not
np.random.seed(13)

random_allocation = np.random.choice(3, 42, p=[0.7, 0.15, 0.15])

np.random.shuffle(random_allocation)

print(random_allocation)
print(Counter(random_allocation))

The data is split accordingly and each committer gets a unique label.

import warnings
warnings.filterwarnings('ignore')

train_set = pd.DataFrame(columns=['message', 'author_email', 'project'])
validate_set = pd.DataFrame(columns=['message', 'author_email', 'project'])
test_set = pd.DataFrame(columns=['message', 'author_email', 'project'])

for i, group_object in enumerate(data.groupby('author_email')):
    group_object[1]['label'] = i
    if random_allocation[i] == 0:
        train_set = pd.concat([train_set, group_object[1]])
    if random_allocation[i] == 1:
        validate_set = pd.concat([validate_set, group_object[1]])
    if random_allocation[i] == 2:
        test_set = pd.concat([test_set, group_object[1]])

train_set.reset_index(drop=True, inplace=True)
validate_set.reset_index(drop=True, inplace=True)
test_set.reset_index(drop=True, inplace=True)

### Check for Overlapping Projects

projects = []
projects.append(list(train_set['project'].unique()))
projects.append(list(validate_set['project'].unique()))
projects.append(list(test_set['project'].unique()))

train_val_overlap = len(set(projects[0]) & set(projects[1]))
train_test_overlap = len(set(projects[0]) & set(projects[2]))
val_test_overlap = len(set(projects[1]) & set(projects[2]))

print(f"There are {train_val_overlap} projects that occur in both train and validate set.")
print(f"There are {train_test_overlap} projects that occur in both train and test set.")
print(f"There are {val_test_overlap} projects that occur in both validate and test set.")

### Resulting Train Set

authors_count = len(train_set['author_email'].unique())
projects_count = len(train_set['project'].unique())

print('Number of different Authors in the train set: ' + str(authors_count))
print('Number of different Projects in the train set: ' + str(projects_count))
print('Average amount of commit messages per author: ' + str(round(len(train_set) / authors_count, 2)))
train_set

### Resulting Validate Set

authors_count = len(validate_set['author_email'].unique())
projects_count = len(validate_set['project'].unique())

print('Number of different Authors in the validate set: ' + str(authors_count))
print('Number of different Projects in the validate set: ' + str(projects_count))
print('Average amount of commit messages per author: ' + str(round(len(validate_set) / authors_count, 2)))
validate_set

### Resulting Test Set

authors_count = len(test_set['author_email'].unique())
projects_count = len(test_set['project'].unique())

print('Number of different Authors in the test set: ' + str(authors_count))
print('Number of different Projects in the test set: ' + str(projects_count))
print('Average amount of commit messages per author: ' + str(round(len(test_set) / authors_count, 2)))
test_set

The number of average commit messages per dataframe cannot be totally balanced since there are some committers with a significantly higher amount of commit messages who are more likely to be allocated to the train set.

### Save all three Dataframes

train_set.to_pickle('../data/04a_Train_Set.pkl')
validate_set.to_pickle('../data/04b_Validate_Set.pkl')
test_set.to_pickle('../data/04c_Test_Set.pkl')

## Split a Second Time

This time the goal is to have commit messages of every author in every dataset to predict authors.

train_set = pd.DataFrame(columns=['message', 'author_email', 'project'])
validate_set = pd.DataFrame(columns=['message', 'author_email', 'project'])
test_set = pd.DataFrame(columns=['message', 'author_email', 'project'])

for i, group_object in enumerate(data.groupby('author_email')):
    group_object[1]['label'] = i
    group_object[1].reset_index(inplace=True)
    random_allocation = np.random.choice(3, len(group_object[1]), p=[0.7, 0.15, 0.15])
    for j, subset in enumerate(random_allocation):
        if subset == 0:
            train_set = pd.concat([train_set, group_object[1].iloc[[j]]])
        if subset == 1:
            validate_set = pd.concat([validate_set, group_object[1].iloc[[j]]])
        if subset == 2:
            test_set = pd.concat([test_set, group_object[1].iloc[[j]]])

train_set.reset_index(drop=True, inplace=True)
validate_set.reset_index(drop=True, inplace=True)
test_set.reset_index(drop=True, inplace=True)

train_set.drop(columns=['index'], inplace=True)
validate_set.drop(columns=['index'], inplace=True)
test_set.drop(columns=['index'], inplace=True)

### Resulting Train Set

authors_count = len(train_set['author_email'].unique())
projects_count = len(train_set['project'].unique())

print('Number of different Authors in the train set: ' + str(authors_count))
print('Number of different Projects in the train set: ' + str(projects_count))
print('Average amount of commit messages per author: ' + str(round(len(train_set) / authors_count, 2)))
train_set

### Resulting Validate Set

authors_count = len(validate_set['author_email'].unique())
projects_count = len(validate_set['project'].unique())

print('Number of different Authors in the validate set: ' + str(authors_count))
print('Number of different Projects in the validate set: ' + str(projects_count))
print('Average amount of commit messages per author: ' + str(round(len(validate_set) / authors_count, 2)))
validate_set

### Resulting Test Set

authors_count = len(test_set['author_email'].unique())
projects_count = len(test_set['project'].unique())

print('Number of different Authors in the test set: ' + str(authors_count))
print('Number of different Projects in the test set: ' + str(projects_count))
print('Average amount of commit messages per author: ' + str(round(len(test_set) / authors_count, 2)))
test_set

### Save all three Dataframes

train_set.to_pickle('../data/05a_Authors_Train_Set.pkl')
validate_set.to_pickle('../data/05b_Authors_Validate_Set.pkl')
test_set.to_pickle('../data/05c_Authors_Test_Set.pkl')
