==> 01_Data Preparation/04_Prebuild_Featuresets.ipynb <==
# Prebuild the featuresets to reduce computational effort

import pandas as pd

train_data = pd.read_pickle('../data/05-0a_Authors_Train_Set.pkl')
test_data = pd.read_pickle('../data/0-05c_Authors_Test_Set.pkl')

import sys
sys.path.append('..')

from util.features import build_featureset

build_featureset(train_data, len(train_data), output_path='../data/05a_Train_Set_Features')
build_featureset(test_data, len(test_data), output_path='../data/05c_Test_Set_Features')

# load data with these commands

import numpy as np

features = np.load('../data/05a_Train_Set_Features.npy', allow_pickle=True)
features = np.load('../data/05c_Test_Set_Features.npy', allow_pickle=True)
